// src/modules/web3/dto/callData.dto.ts
import { MintDataDto } from './mintData.dto';
import { Networks, OperationTypes } from '../../../common/constants';
import { WhitelistDto } from './whitelist.dto';

/**
 * A data transfer object for passing call data.
 */
export class CallDataDto {
  execute: boolean;
  network?: Networks;
  contract_id: string;
  method_name: string;
  arguments: string | null;
  from_address: string;
  operation_type?: OperationTypes;
  operation_options?: MintDataDto | WhitelistDto;
}